Prepare 06 based on Eric Vyncke AD review
-	Generally, a justification of why using DHCP for MAC allocation beyond the one paragraph in section 1 would be welcome. 
BV> There’s also the end of the introduction that further motivates this. Is that not sufficient or should the section be reorganized a bit.

-	Section 1, in “form of local authority” is ‘local’ really required in this context ?
BV> I will remove local.

-	Section 1, nice to read that the mechanism could be extended beyond the usual 48-bit MAC address

-	Section 3, the “address assignment » should probably be qualified as, in the rest of this document, addresses are MAC addresses (see also 2 paragraphs below!)
BV> I think the references you are referring to are in “The client may or may not support address assignment”; can change to “IPv6 address assignment”.

-	Nit but sometime IEEE802 is written as IEEE.802 or IEEE 802 => suggest to be consistent
BV> Using “IEEE 802”.

-	Section 4.1, “unexpected for all MAC addresses to be released” is the local pool exhaustion a problem? Suggest to either do not state this or mention the section 4.3 and the ‘release’ operation
BV> Removed and reword slightly.

-	Section 4.2 suggest to expand DUID on first use (even if quite obvious for a DHC document)
BV> OK, reworked sentence to allow for this: Note that a client that operates as above that does not have a globally unique link-layer address on any of its interfaces MUST NOT use a link-layer based DUID (DHCP Unique Identifier), i.e., DUID-LLT or DUID-LL.

-	Last paragraph of section 4.3 is “DISCUSSION”... is it a left over of the WGLC ?
BV> Removed paragraph; already covered in first that client and request a specific address.

-	There is a mix of DHCP and DHCPv6, suggest to stick to DHCP only
BV> I reduced the DHCPv6 usage; I left DHCPv6 in the Abstract and Introduction.

-	In the same vein, there is a mix of address, MAC address, link-layer address..., suggest to stick to “address” per the terminology section
BV> I greatly reduced the MAC/link-layer before address, but left some to assure proper context as just saying “address” at times can be ambiguous.

-	Section 8, the “SHOULD NOT” does not live well together with the following “MUST NOT”, may be I am wrong but this should be all “SHOULD NOT” or “MUST NOT” but not a mix
BV> I removed the “SHOULD NOT” sentence as it really wasn’t needed.

-	Section 10.2, specify the unit for link-layer-len (even if obvious)
BV> Added in octets - “Specifies the length, in octets, …”

-	Possibly naïve, but, should the user of a leased MAC address perform some kind of DAD ?
BV> In section 7, we have:
   A client that changes its link-layer address on an interface SHOULD
   follow the recommendations in Section 7.2.6 of [RFC4861] to inform
   its neighbors of the new link-layer address quickly.
I think that’s about as good as we have?  Note that the node may or may  not have IPv6 addresses … are you suggesting we should also require DAD for the IPv6 addresses?